codegen/requirements.txt
libclang==16.0.6
Jinja2==3.1.2
PyYAML==6.0